refactor: split topics screen into components and hooks
