+++ app-expo/features/topics/TopicCard.tsx
@@ -1,0 +1,103 @@
+// TopicCard - displays a single topic with image and hide action
+import React from 'react';
+import { View, Text, Image, TouchableOpacity, StyleSheet, Dimensions } from 'react-native';
+import { Trash } from 'lucide-react-native';
+import { Topic } from '@/types/search';
+
+type Props = {
+  topic: Topic;
+  onHide: (id: string) => void;
+};
+
+const { width, height } = Dimensions.get('window');
+const CARD_WIDTH = width - 32;
+const CARD_HEIGHT = height * 0.85;
+
+export function TopicCard({ topic, onHide }: Props) {
+  return (
+    <View style={styles.card}>
+      <Image source={{ uri: topic.imageUrl }} style={styles.cardImage} />
+      <View style={styles.cardOverlay}>
+        <TouchableOpacity style={styles.hideButton} onPress={() => onHide(topic.id)}>
+          <Trash size={18} color="#FFF" />
+        </TouchableOpacity>
+        <View style={styles.cardContent}>
+          <Text style={styles.cardTitle}>{topic.topicTitle}</Text>
+          <Text style={styles.cardDescription}>{topic.reason}</Text>
+        </View>
+      </View>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  card: {
+    width: CARD_WIDTH,
+    height: CARD_HEIGHT,
+    borderRadius: 24,
+    overflow: 'hidden',
+    borderWidth: 4,
+    borderColor: '#fff',
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 0 },
+    shadowOpacity: 0.3,
+    shadowRadius: 32,
+    elevation: 12,
+    position: 'relative',
+  },
+  cardImage: {
+    width: '100%',
+    height: '100%',
+    resizeMode: 'cover',
+  },
+  cardOverlay: {
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    right: 0,
+    bottom: 0,
+    backgroundColor: 'rgba(0, 0, 0, 0.1)',
+    padding: 24,
+    justifyContent: 'space-between',
+  },
+  hideButton: {
+    alignSelf: 'flex-end',
+    flexDirection: 'row',
+    alignItems: 'center',
+    backgroundColor: 'rgba(0, 0, 0, 0.3)',
+    paddingHorizontal: 16,
+    paddingVertical: 10,
+    borderRadius: 20,
+    gap: 6,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.3,
+    shadowRadius: 4,
+    elevation: 4,
+  },
+  cardContent: {
+    flex: 1,
+    justifyContent: 'flex-end',
+  },
+  cardTitle: {
+    fontSize: 24,
+    color: '#FFF',
+    fontWeight: '700',
+    lineHeight: 32,
+    letterSpacing: -0.3,
+    marginBottom: 12,
+    textShadowColor: 'rgba(0, 0, 0, 0.8)',
+    textShadowOffset: { width: 0, height: 1 },
+    textShadowRadius: 3,
+  },
+  cardDescription: {
+    fontSize: 16,
+    color: '#E0E0E0',
+    lineHeight: 24,
+    marginBottom: 16,
+    textShadowColor: 'rgba(0, 0, 0, 0.8)',
+    textShadowOffset: { width: 0, height: 1 },
+    textShadowRadius: 3,
+    fontWeight: '500',
+  },
+});
